Move menu factory to apputils
